Modified lynxnet2
A modified lynxnet2.py file to fix an issue regarding variance training 

Thank you MLo7Ghinsan for helping me figuring out the toplevel.py and for providing the lynxnet2.py file!

+++ modules/backbones/lynxnet2.py
@@ -90,7 +90,7 @@ class LYNXNet2(nn.Module):
         else:
             x = spec.flatten(start_dim=1, end_dim=2)  # [B, F x M, T]
 
-        x = self.input_projection(x.transpose(1, 2)) # [B, T, F x M]
+        x = self.input_projection(x.transpose(1, 2).float()) # [B, T, F x M]
         if self.use_conditioner_cache:
             # It may need to be modified at some point to be compatible with the condition cache
             x = x + self.conditioner_projection(cond).transpose(1, 2)
